Split registry shutdown code into seperate module
The main module should only be daemon code, not also remote registry
shutdown code.
